+++ data/BMAssetMap.js
@@ -1,0 +1,121 @@
+// BM asset path map for future shop/BM UI rebuild.
+// This file is intentionally not imported yet. Use it when replacing the current shop UI structure.
+
+export const BM_ASSETS = Object.freeze({
+  buttons: Object.freeze({
+    confirm: "./assets/images/bm/buttons/base/btn_confirm.png",
+    cancel: "./assets/images/bm/buttons/base/btn_cancel.png",
+    buy: "./assets/images/bm/buttons/base/btn_buy.png",
+    buyDisabled: "./assets/images/bm/buttons/base/btn_buy_disabled.png",
+    contract: "./assets/images/bm/buttons/base/btn_contract.png",
+    skip: "./assets/images/bm/buttons/base/btn_skip.png",
+    close: "./assets/images/bm/buttons/extra/btn_close.png",
+    use: "./assets/images/bm/buttons/extra/btn_use.png",
+    upgrade: "./assets/images/bm/buttons/extra/btn_upgrade.png",
+    claim: "./assets/images/bm/buttons/extra/btn_claim.png",
+    watchAd: "./assets/images/bm/buttons/extra/btn_watch_ad.png"
+  }),
+  currency: Object.freeze({
+    gold: "./assets/images/bm/icons/currency/currency_gold.png",
+    diamond: "./assets/images/bm/icons/currency/currency_diamond.png"
+  }),
+  rewardIcons: Object.freeze({
+    adReward: "./assets/images/bm/icons/reward/icon_ad_reward.png",
+    reward: "./assets/images/bm/icons/reward/icon_reward.png"
+  }),
+  statusIcons: Object.freeze({
+    locked: "./assets/images/bm/icons/status/icon_locked.png",
+    unlockable: "./assets/images/bm/icons/status/icon_unlockable.png",
+    contractable: "./assets/images/bm/icons/status/icon_contractable.png",
+    onSale: "./assets/images/bm/icons/status/icon_on_sale.png",
+    notReady: "./assets/images/bm/icons/status/icon_not_ready.png"
+  }),
+  items: Object.freeze({
+    skip3Day: "./assets/images/bm/items/item_skip_3day.png",
+    instantExpansion: "./assets/images/bm/items/item_instant_expansion.png",
+    adSkipTicket: "./assets/images/bm/items/item_ad_skip_ticket.png"
+  }),
+  coupons: Object.freeze({
+    adSkip: "./assets/images/bm/coupons/coupon_ad_skip.png",
+    peakTime: "./assets/images/bm/coupons/coupon_peak_time.png"
+  }),
+  labels: Object.freeze({
+    priceBase: "./assets/images/bm/labels/label_price_base.png",
+    priceGold: "./assets/images/bm/labels/label_price_gold.png",
+    priceDiamond: "./assets/images/bm/labels/label_price_diamond.png",
+    priceCash: "./assets/images/bm/labels/label_price_cash.png",
+    free: "./assets/images/bm/labels/label_free.png",
+    owned: "./assets/images/bm/labels/label_owned.png",
+    soldOut: "./assets/images/bm/labels/label_sold_out.png",
+    dailyLimit: "./assets/images/bm/labels/label_daily_limit.png",
+    discount05: "./assets/images/bm/labels/label_discount_05.png",
+    discount10: "./assets/images/bm/labels/label_discount_10.png",
+    discount15: "./assets/images/bm/labels/label_discount_15.png",
+    discount50: "./assets/images/bm/labels/label_discount_50.png"
+  }),
+  popups: Object.freeze({
+    base: "./assets/images/bm/popups/popup_base.png",
+    productUnlock: "./assets/images/bm/popups/popup_product_unlock.png",
+    productContract: "./assets/images/bm/popups/popup_product_contract.png",
+    skipTicketUse: "./assets/images/bm/popups/popup_skip_ticket_use.png",
+    purchaseComplete: "./assets/images/bm/popups/popup_purchase_complete.png",
+    currencyShortage: "./assets/images/bm/popups/popup_currency_shortage.png",
+    instantExpansionUse: "./assets/images/bm/popups/popup_instant_expansion_use.png",
+    peakCouponPurchase: "./assets/images/bm/popups/popup_peak_coupon_purchase.png",
+    peakCouponUse: "./assets/images/bm/popups/popup_peak_coupon_use.png",
+    adSkipUse: "./assets/images/bm/popups/popup_ad_skip_use.png",
+    dailyMissionReward: "./assets/images/bm/popups/popup_daily_mission_reward.png",
+    productUpgrade: "./assets/images/bm/popups/popup_product_upgrade.png",
+    shelfUpgrade: "./assets/images/bm/popups/popup_shelf_upgrade.png",
+    storageExpansion: "./assets/images/bm/popups/popup_storage_expansion.png",
+    staffHire: "./assets/images/bm/popups/popup_staff_hire.png",
+    staffUpgrade: "./assets/images/bm/popups/popup_staff_upgrade.png",
+    storageFull: "./assets/images/bm/popups/popup_storage_full.png"
+  }),
+  shop: Object.freeze({
+    bg: "./assets/images/bm/shop/ui_shop_bg.png",
+    topBar: "./assets/images/bm/shop/ui_shop_top_bar.png",
+    cardBase: "./assets/images/bm/shop/ui_shop_card_base.png",
+    cardLocked: "./assets/images/bm/shop/ui_shop_card_locked.png",
+    cardDisabled: "./assets/images/bm/shop/ui_shop_card_disabled.png",
+    cardPremium: "./assets/images/bm/shop/ui_shop_card_premium.png",
+    tabs: Object.freeze({
+      currencyDefault: "./assets/images/bm/shop/tab_currency_default.png",
+      currencySelected: "./assets/images/bm/shop/tab_currency_selected.png",
+      itemDefault: "./assets/images/bm/shop/tab_item_default.png",
+      itemSelected: "./assets/images/bm/shop/tab_item_selected.png",
+      premiumDefault: "./assets/images/bm/shop/tab_premium_default.png",
+      premiumSelected: "./assets/images/bm/shop/tab_premium_selected.png"
+    })
+  }),
+  productList: Object.freeze({
+    bg: "./assets/images/bm/product_list/ui_product_list_bg.png",
+    thumbFrame: "./assets/images/bm/product_list/ui_product_thumb_frame.png",
+    detailPanel: "./assets/images/bm/product_list/ui_product_detail_panel.png",
+    cardBase: "./assets/images/bm/product_list/ui_product_card_base.png",
+    cardLocked: "./assets/images/bm/product_list/ui_product_card_locked.png",
+    cardUnlockable: "./assets/images/bm/product_list/ui_product_card_unlockable.png",
+    cardContractable: "./assets/images/bm/product_list/ui_product_card_contractable.png",
+    cardOwned: "./assets/images/bm/product_list/ui_product_card_owned.png",
+    cardOnSale: "./assets/images/bm/product_list/ui_product_card_on_sale.png",
+    cardNotReady: "./assets/images/bm/product_list/ui_product_card_not_ready.png",
+    tabs: Object.freeze({
+      basicShelf: "./assets/images/bm/product_list/tab_category_basic_shelf.png",
+      freshShelf: "./assets/images/bm/product_list/tab_category_fresh_shelf.png",
+      fridge: "./assets/images/bm/product_list/tab_category_fridge.png",
+      warmer: "./assets/images/bm/product_list/tab_category_warmer.png"
+    })
+  }),
+  freeCharge: Object.freeze({
+    panel: "./assets/images/bm/free_charge/ui_free_charge_panel.png",
+    cardBase: "./assets/images/bm/free_charge/ui_ad_reward_card_base.png",
+    cardComplete: "./assets/images/bm/free_charge/ui_ad_reward_card_complete.png",
+    diamondRewardIcon: "./assets/images/bm/free_charge/icon_ad_diamond_reward.png",
+    goldRewardIcon: "./assets/images/bm/free_charge/icon_ad_gold_reward.png",
+    peakCouponDiscountIcon: "./assets/images/bm/free_charge/icon_ad_peak_coupon_discount.png"
+  })
+});
+
+export function getBmAssetPath(...keys) {
+  return keys.reduce((current, key) => current?.[key], BM_ASSETS) ?? null;
+}